Move over rest of iso calculations
